requirements.txt
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==20.1.0
astroid==2.4.2
async-generator==1.10
attrs==20.3.0
autopep8==1.5.6
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.9.3
black==20.8b1
bleach==3.2.1
blinker==1.4
boto3==1.17.73
botocore==1.20.73
Brlapi==0.7.0
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.3
command-not-found==0.3
contextlib2==0.6.0.post1
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==2.2.20
djangorestframework==3.12.4
docopt==0.6.2
duplicity==0.8.12.0
dynaconf==3.1.4
dython==0.6.2
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
httplib2==0.14.0
hvac==0.10.11
idna==2.8
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.2
isort==5.6.4
jedi==0.18.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==1.0.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keyring==18.0.1
kiwisolver==1.3.1
knox==0.1.11
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
loguru==0.5.3
louis==3.12.0
lxml==4.6.2
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.3.3
mccabe==0.6.1
mistune==0.8.4
monotonic==1.5
mypy-extensions==0.4.3
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
netifaces==0.10.4
notebook==6.1.6
numpy==1.19.4
oauthlib==3.1.0
olefile==0.46
packaging==20.8
pandas==1.2.0
pandocfilters==1.4.3
paramiko==2.6.0
parso==0.8.1
pathspec==0.8.1
pbr==5.5.1
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.6.1
ptyprocess==0.7.0
pycairo==1.16.2
pycodestyle==2.7.0
pycparser==2.20
pycups==1.9.73
pydot==1.4.1
Pygments==2.7.3
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.17.3
PySocks==1.7.1
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.15.0
pythonanywhere==0.9.6
PythonTurtle==0.3.2
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
regex==2020.11.13
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
s3transfer==0.4.2
schema==0.7.4
scikit-learn==0.24.0
scipy==1.6.0
screen-resolution-extra==0.0.0
seaborn==0.11.1
SecretStorage==2.3.1
selenium==3.141.0
Send2Trash==1.5.0
shortuuid==1.0.1
simplejson==3.16.0
six==1.14.0
soupsieve==2.1
sqlparse==0.4.1
stevedore==3.3.0
systemd-python==234
tabulate==0.8.9
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
tweepy==3.10.0
typed-ast==1.4.2
typer==0.3.2
typing-extensions==3.7.4.3
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
validators==0.18.2
virtualenv==20.4.4
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xkit==0.0.0
